# core-lending-regulatory-compliance-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-regulatory-compliance
  description: Lending Regulatory & Compliance Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-compliance
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/reporting-runs/{reportingRunId}:
    get:
      tags:
        - ReportingRun
      summary: Get a reporting run by ID
      operationId: getById
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReportingRunDTO'
    put:
      tags:
        - ReportingRun
      summary: Update a reporting run
      operationId: update
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportingRunDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReportingRunDTO'
    delete:
      tags:
        - ReportingRun
      summary: Delete a reporting run
      operationId: delete
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/reporting-runs/{reportingRunId}/records/{reportingRecordId}:
    get:
      tags:
        - ReportingRecord
      summary: Get a reporting record by ID
      operationId: getById_2
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: reportingRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReportingRecordDTO'
    put:
      tags:
        - ReportingRecord
      summary: Update a reporting record
      operationId: update_2
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: reportingRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportingRecordDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReportingRecordDTO'
    delete:
      tags:
        - ReportingRecord
      summary: Delete a reporting record
      operationId: delete_2
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: reportingRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/reporting-runs:
    get:
      tags:
        - ReportingRun
      summary: List or search reporting runs
      operationId: findAll
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.reportType
          in: query
          description: Exact filter for report type
          required: false
          schema:
            type: string
            enum:
              - CIRBE
              - FINREP
              - AML
              - COREP
              - SEPBLAC
        - name: filters.reportingPeriodStart
          in: query
          description: Exact filter for reporting period start
          required: false
          schema:
            type: string
        - name: filters.reportingPeriodEnd
          in: query
          description: Exact filter for reporting period end
          required: false
          schema:
            type: string
        - name: filters.runStatus
          in: query
          description: Exact filter for run status
          required: false
          schema:
            type: string
            enum:
              - OPEN
              - IN_PROGRESS
              - COMPLETED
              - FAILED
              - CANCELLED
        - name: filters.remarks
          in: query
          description: Exact filter for remarks
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseReportingRunDTO'
    post:
      tags:
        - ReportingRun
      summary: Create a new reporting run
      operationId: create
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportingRunDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReportingRunDTO'
...
  /api/v1/reporting-runs/{reportingRunId}/records:
    get:
      tags:
        - ReportingRecord
      summary: List or search reporting records for a specific run
      operationId: findAll_2
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.reportingRunId
          in: query
          description: Exact filter for reporting run id
          required: false
          schema:
            type: string
        - name: filters.loanServicingCaseId
          in: query
          description: Exact filter for loan servicing case id
          required: false
          schema:
            type: string
        - name: filters.customerId
          in: query
          description: Exact filter for customer id
          required: false
          schema:
            type: string
        - name: filters.transactionId
          in: query
          description: Exact filter for transaction id
          required: false
          schema:
            type: string
        - name: filters.recordType
          in: query
          description: Exact filter for record type
          required: false
          schema:
            type: string
            enum:
              - EXPOSURE
              - DELINQUENCY
              - AML_TXN
              - COLLATERAL
              - PROVISION
        - name: filters.reportedAmount
          in: query
          description: Exact filter for reported amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[reportedAmount].from
          in: query
          description: Filter reported amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[reportedAmount].to
          in: query
          description: Filter reported amount to value
          required: false
          schema:
            type: string
        - name: filters.currencyCode
          in: query
          description: Exact filter for currency code
          required: false
          schema:
            type: string
            enum:
              - EUR
              - USD
              - GBP
              - CHF
        - name: filters.additionalData
          in: query
          description: Exact filter for additional data
          required: false
          schema:
            type: string
        - name: filters.isIncluded
          in: query
          description: Exact filter for is included
          required: false
          schema:
            type: boolean
        - name: filters.exclusionReason
          in: query
          description: Exact filter for exclusion reason
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseReportingRecordDTO'
    post:
      tags:
        - ReportingRecord
      summary: Create a new reporting record
      operationId: create_2
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportingRecordDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReportingRecordDTO'
...
components:
  schemas:
    ReportingRunDTO:
      required:
        - reportType
        - reportingPeriodEnd
        - reportingPeriodStart
        - runStatus
      type: object
      properties:
        reportingRunId:
          type: string
          format: uuid
          readOnly: true
        reportType:
          type: string
          enum:
            - CIRBE
            - FINREP
            - AML
            - COREP
            - SEPBLAC
        reportingPeriodStart:
          type: string
          format: date
        reportingPeriodEnd:
          type: string
          format: date
        runStatus:
          type: string
          enum:
            - OPEN
            - IN_PROGRESS
            - COMPLETED
            - FAILED
            - CANCELLED
        remarks:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ReportingRecordDTO:
      required:
        - currencyCode
        - customerId
        - isIncluded
        - loanServicingCaseId
        - recordType
        - reportedAmount
        - reportingRunId
        - transactionId
      type: object
      properties:
        reportingRecordId:
          type: string
          format: uuid
          readOnly: true
        reportingRunId:
          type: string
          format: uuid
        loanServicingCaseId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        recordType:
          type: string
          enum:
            - EXPOSURE
            - DELINQUENCY
            - AML_TXN
            - COLLATERAL
            - PROVISION
        reportedAmount:
          minimum: 0
          exclusiveMinimum: true
          type: number
        currencyCode:
          type: string
          enum:
            - EUR
            - USD
            - GBP
            - CHF
        additionalData:
          maxLength: 5000
          minLength: 0
          type: string
        isIncluded:
          type: boolean
        exclusionReason:
          maxLength: 500
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseReportingRunDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ReportingRunDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    PaginationResponseReportingRecordDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ReportingRecordDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...