# core-lending-regulatory-compliance-web/src/main/java/com/firefly/core/lending/compliance/web/controllers/reporting/v1/ReportingRecordController.java
/*
 * Copyright 2025 Firefly Software Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.firefly.core.lending.compliance.web.controllers.reporting.v1;

import org.fireflyframework.core.filters.FilterRequest;
import org.fireflyframework.core.queries.PaginationResponse;
import com.firefly.core.lending.compliance.core.services.reporting.v1.ReportingRecordService;
import com.firefly.core.lending.compliance.interfaces.dtos.reporting.v1.ReportingRecordDTO;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import reactor.core.publisher.Mono;

import java.util.UUID;

@RestController
@RequestMapping("/api/v1/reporting-runs/{reportingRunId}/records")
@Tag(name = "ReportingRecord", description = "Operations for Reporting Records within a Reporting Run")
@RequiredArgsConstructor
public class ReportingRecordController {

    private final ReportingRecordService service;

    @GetMapping
    @Operation(summary = "List or search reporting records for a specific run")
    public Mono<ResponseEntity<PaginationResponse<ReportingRecordDTO>>> findAll(
            @PathVariable UUID reportingRunId,
            @ModelAttribute FilterRequest<ReportingRecordDTO> filterRequest) {

        return service.findAll(reportingRunId, filterRequest)
                .map(ResponseEntity::ok);
    }

    @PostMapping
    @Operation(summary = "Create a new reporting record")
    public Mono<ResponseEntity<ReportingRecordDTO>> create(
            @PathVariable UUID reportingRunId,
            @RequestBody ReportingRecordDTO dto) {

        return service.create(reportingRunId, dto)
                .map(ResponseEntity::ok);
    }

    @GetMapping("/{reportingRecordId}")
    @Operation(summary = "Get a reporting record by ID")
    public Mono<ResponseEntity<ReportingRecordDTO>> getById(
            @PathVariable UUID reportingRunId,
            @PathVariable UUID reportingRecordId) {

        return service.getById(reportingRunId, reportingRecordId)
                .map(ResponseEntity::ok);
    }

    @PutMapping("/{reportingRecordId}")
    @Operation(summary = "Update a reporting record")
    public Mono<ResponseEntity<ReportingRecordDTO>> update(
            @PathVariable UUID reportingRunId,
            @PathVariable UUID reportingRecordId,
            @RequestBody ReportingRecordDTO dto) {

        return service.update(reportingRunId, reportingRecordId, dto)
                .map(ResponseEntity::ok);
    }

    @DeleteMapping("/{reportingRecordId}")
    @Operation(summary = "Delete a reporting record")
    public Mono<ResponseEntity<Void>> delete(
            @PathVariable UUID reportingRunId,
            @PathVariable UUID reportingRecordId) {

        return service.delete(reportingRunId, reportingRecordId)
                .thenReturn(ResponseEntity.noContent().build());
    }
}